Fixed two bugs related to speculative reading for singlet
The speculative writes were not correctly picked for Singlet, it was rejecting same
singlet instead of only accepting them! At least we always had a single Singlet in
StateDB implementation on EOSIO, so never got hit by this bug.

The speculative entry was not used correctly to determine if it should be considered or
not. We were wrongly using the entry from the database to check if the speculative entry
should have been applied. This was causing a nil pointer when the entry was not present
in the database since it was trying to extract the singlet from a nil entry.