Added possibility to re-index all existing tablet indexes
This can be used up to a certain height, it's highly optimized by re-writting indexes
from low block to high block, scanning only rows in-between each by leveraging tablet
cache.

Fixed some possible issues on Ethereum where some exclusive bounds where on 0 where it
could be a problem on Ethereum is it was excluded (it should be included).