Fixed shard injector
The shard injector was fetching last written checkpoint and when not there, it was returning `bstream.EmptyBlockRef`. However, we were not checking this condition properly and were blindly starting after last checkpoint block's num. Since the empty block ref block's num is 0, we were actually skipping block 0 from being written to databse.

When injecting shard now, last checkpoint is now check if it's the empty block ref and if it's the case, we do not check if we should start later or not.