Fixed `ReadSingletEntryAt` when trying to query a singlet entry at height 0 directly
The `endKey` is exlusive but when querying for specifically height 0, we were creating a scan of the form `start: ffff, exlusiveEnd: ffff` (`ffff` because we use reverse height notation), since `exclusiveEnd: ffff` is exlusive and as such, `ffff` value was excluded.

To fix this, we now use `kvdb/store.Key(<key>).Next()` to compute the actual last exclusive key to query up to.